perf: 修改数据库结构
去掉了外键约束
使用逻辑删除
索引和查询优化

+++ services/core/src/main/java/cloud/oj/core/service/ContestService.java
@@ -30,40 +30,47 @@ public class ContestService {
     }
 
     public Optional<Contest> getContest(int contestId) {
-        return Optional.ofNullable(contestDao.getContest(contestId));
+        return Optional.ofNullable(contestDao.getContestById(contestId));
     }
 
     public List<List<?>> getStartedContest(int page, int limit) {
         return contestDao.getContests((page - 1) * limit, limit, true);
     }
 
-    public Integer addContest(Contest contest) {
+    public HttpStatus addContest(Contest contest) {
         if (contestDao.addContest(contest) == 1) {
-            return HttpStatus.CREATED.value();
+            return HttpStatus.CREATED;
         } else {
-            throw new GenericException(HttpStatus.BAD_REQUEST.value(), "请求数据可能不正确");
+            throw new GenericException(400, "请求数据可能不正确");
         }
     }
 
-    public Integer updateContest(Contest contest) {
+    public HttpStatus updateContest(Contest contest) {
         if (contestDao.updateContest(contest) == 1) {
-            return HttpStatus.OK.value();
+            return HttpStatus.OK;
         } else {
-            throw new GenericException(
-                    HttpStatus.NOT_MODIFIED.value(),
-                    String.format("竞赛 %d 更新失败", contest.getContestId())
-            );
+            throw new GenericException(400, String.format("竞赛(%d)更新失败", contest.getContestId()));
         }
     }
 
-    public Integer deleteContest(int contestId) {
+    /**
+     * 逻辑删除竞赛
+     * <p>已开始未结束的竞赛不允许删除</p>
+     *
+     * @param contestId 竞赛 Id
+     * @return HTTP 状态码 {@link HttpStatus}
+     */
+    public HttpStatus deleteContest(int contestId) {
+        var contest = contestDao.getState(contestId);
+
+        if (contest.isStarted() && !contest.isEnded()) {
+            throw new GenericException(400, "竞赛已开始，不准删除");
+        }
+
         if (contestDao.deleteContest(contestId) == 1) {
-            return HttpStatus.NO_CONTENT.value();
+            return HttpStatus.NO_CONTENT;
         } else {
-            throw new GenericException(
-                    HttpStatus.GONE.value(),
-                    String.format("竞赛 %d 不存在", contestId)
-            );
+            throw new GenericException(410, String.format("竞赛(%d)不存在", contestId));
         }
     }
 
@@ -72,26 +79,30 @@ public class ContestService {
     }
 
     public List<Problem> getProblemsFromContest(String userId, int contestId, boolean onlyStarted) {
-        return contestDao.getProblems(userId, contestId, onlyStarted);
+        if (userId == null) {
+            return contestDao.getProblems(contestId, onlyStarted);
+        }
+
+        return contestDao.getProblemsWithResult(userId, contestId);
     }
 
     public Optional<Problem> getProblemInContest(Integer contestId, Integer problemId) {
-        return Optional.ofNullable(problemDao.getSingleInContest(contestId, problemId));
+        return Optional.ofNullable(problemDao.getByIdFromContest(contestId, problemId));
     }
 
-    public Integer addProblemToContest(int contestId, int problemId) {
+    public HttpStatus addProblemToContest(int contestId, int problemId) {
         if (contestDao.addProblem(contestId, problemId) == 1) {
-            return HttpStatus.CREATED.value();
+            return HttpStatus.CREATED;
         } else {
-            throw new GenericException(HttpStatus.BAD_REQUEST.value(), "无法添加题目");
+            throw new GenericException(400, "无法添加题目");
         }
     }
 
-    public Integer removeProblem(int contestId, int problemId) {
+    public HttpStatus removeProblem(int contestId, int problemId) {
         if (contestDao.removeProblem(contestId, problemId) == 1) {
-            return HttpStatus.NO_CONTENT.value();
+            return HttpStatus.NO_CONTENT;
         } else {
-            throw new GenericException(HttpStatus.GONE.value(), String.format("题目 %d 不存在", problemId));
+            throw new GenericException(410, String.format("题目(%d)不存在", problemId));
         }
     }
 }
